src/pyproject.toml
[tool.poetry]
name = "my virtualenv"
version = "0.1.0"
description = ""
authors = ["andrey <a.arbatskov@purpleplane.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
click = "^8.1.3"
strawberry-graphql = {extras = ["debug-server"], version = "^0.137.1"}
sqlalchemy = "^1.4.42"
alembic = "^1.8.1"
psycopg2 = "^2.9.5"
asyncpg = "^0.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
